move EngineType to hudi-client-common modeule
